Refactoring to r8 app module, cleanup and adjustments
