+++ example/androidlib/java/5-R8/build.mill
@@ -24,7 +24,7 @@ object androidSdkModule0 extends AndroidSdkModule { // <1>
   def buildToolsVersion = "35.0.0"
 }
 
-object app extends AndroidAppModule { // <2>
+object app extends AndroidR8AppModule { // <2>
   def androidSdkModule = mill.define.ModuleRef(androidSdkModule0)
   def androidMinSdk = 19
   def androidCompileSdk = 35
@@ -61,7 +61,8 @@ object app extends AndroidAppModule { // <2>
   }
 
   // Instrumented tests (runs on emulator)
-  object it extends AndroidAppInstrumentedTests with AndroidTestModule.AndroidJUnit {
+  object it extends AndroidAppInstrumentedTests with AndroidR8AppModule
+      with AndroidTestModule.AndroidJUnit {
     def androidSdkModule = mill.define.ModuleRef(androidSdkModule0)
 
     override def androidIsDebug: T[Boolean] = Task {
@@ -69,13 +70,13 @@ object app extends AndroidAppModule { // <2>
     }
 
     override def androidReleaseSettings: T[AndroidBuildTypeSettings] = Task {
-      AndroidBuildTypeSettings(isMinifyEnabled = false)
+      AndroidBuildTypeSettings(isMinifyEnabled = false).withProguardLocalFiles(
+        Seq(
+          moduleDir / "test-proguard-rules.pro"
+        )
+      )
     }
 
-    /* TODO currently the dependency resolution ignores the platform type and kotlinx-coroutines-core has
-     * conflicting classes with kotlinx-coroutines-core-jvm . Remove the exclusions once the dependency
-     * resolution resolves conflicts between androidJvm and jvm platform types
-     */
     def mvnDeps = super.mvnDeps() ++ Seq(
       mvn"androidx.test.ext:junit:1.2.1",
       mvn"androidx.test:runner:1.6.2",
