+++ libs/androidlib/src/mill/androidlib/AndroidAppModule.scala
@@ -806,72 +806,6 @@ trait AndroidAppModule extends AndroidModule { outer =>
       throw new Exception("Device failed to boot")
   }
 
-  def androidLibraryProguardConfigs: Task[Seq[PathRef]] = Task {
-    androidUnpackArchives()
-      // TODO need also collect rules from other modules,
-      // but Android lib module doesn't yet exist
-      .flatMap(_.proguardRules)
-  }
-
-  /** ProGuard/R8 rules configuration files for release target (user-provided and generated) */
-  def androidProguardConfigs: Task[Seq[PathRef]] = Task {
-    val proguardFilesFromBuildSettings = androidBuildSettings().proguardFiles
-    val androidProguardPath = androidSdkModule().androidProguardPath().path
-    val defaultProguardFile = proguardFilesFromBuildSettings.defaultProguardFile.map {
-      pf => androidProguardPath / pf
-    }
-    val userProguardFiles = proguardFilesFromBuildSettings.localFiles
-    mill.define.BuildCtx.withFilesystemCheckerDisabled {
-      (defaultProguardFile.toSeq ++ userProguardFiles).map(PathRef(
-        _
-      )) ++ androidLibraryProguardConfigs()
-    }
-  }
-
-  /** Concatenates all rules into one file */
-  def androidProguard: T[PathRef] = Task {
-    val globalProguard = Task.dest / "global-proguard.pro"
-    val files = androidProguardConfigs()
-    os.write(globalProguard, "")
-    files.foreach(pg =>
-      os.write.append(globalProguard, os.read(pg.path))
-    )
-    PathRef(globalProguard)
-  }
-
-  /**
-   * The default release settings with the following settings:
-   * - minifyEnabled=true
-   * - shrinkEnabled=true
-   * - proguardFiles=proguard-android-optimize.txt
-   * @return
-   */
-  def androidReleaseSettings: T[AndroidBuildTypeSettings] = Task {
-    AndroidBuildTypeSettings(
-      isMinifyEnabled = true,
-      isShrinkEnabled = true,
-      proguardFiles = ProguardFiles(
-        defaultProguardFile = Some("proguard-android-optimize.txt")
-      )
-    )
-  }
-
-  def androidDebugSettings: T[AndroidBuildTypeSettings] = Task {
-    AndroidBuildTypeSettings()
-  }
-
-  /**
-   * Gives the android build type settings for debug or release.
-   * Controlled by [[androidIsDebug]] flag!
-   * @return
-   */
-  def androidBuildSettings: T[AndroidBuildTypeSettings] = Task {
-    if (androidIsDebug())
-      androidDebugSettings()
-    else
-      androidReleaseSettings()
-  }
-
   /**
    * Converts the generated JAR file into a DEX file using the `d8`
    * through the [[androidBuildSettings]].
